pyproject.toml
[project]
name = "resume"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "HanRix1", email = "HanRix1@mail.ru"},
]
dependencies = [
    "fastapi[standard]>=0.115.4",
    "Requests>=2.32.3",
    "pydantic-settings>=2.6.1",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2>=2.9.10",
    "python-multipart>=0.0.17",
    "boto3[crt]>=1.35.56",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
